Backup/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2
pytest-asyncio==0.23.7
httpx==0.27.0
psutil==5.9.8

# JWT Authentication 身份验证系统
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database dependencies (optional)
ibm-db>=3.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

#Agents 
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# macOS 原生框架绑定 用于获取GPS定位
pyobjc==11.1    